+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Transaction hash as starting point.
 - Block hash as starting point.
 - Timestamp as starting point.
+- Configurable xpub gap limit.
 
 ### Removed
 - Mempool endpoint now has no limits or offsets and always returns full list.